+++ crates/zng-view/src/surface.rs
@@ -20,7 +20,8 @@ use crate::{
     display_list::{display_list_to_webrender, DisplayListCache},
     extensions::{
         self, BlobExtensionsImgHandler, DisplayListExtAdapter, FrameReadyArgs, RedrawArgs, RendererCommandArgs, RendererConfigArgs,
-        RendererDeinitedArgs, RendererExtension, RendererInitedArgs,
+        RendererDeinitedArgs, RendererExtension, RendererInitedArgs, WindowCommandArgs, WindowConfigArgs, WindowDeinitedArgs,
+        WindowExtension, WindowInitedArgs,
     },
     gl::{GlContext, GlContextManager},
     image_cache::{Image, ImageCache, ImageUseMap, WrImageCache},
@@ -40,6 +41,7 @@ pub(crate) struct Surface {
 
     context: GlContext,
     renderer: Option<Renderer>,
+    window_exts: Vec<(ApiExtensionId, Box<dyn WindowExtension>)>,
     renderer_exts: Vec<(ApiExtensionId, Box<dyn RendererExtension>)>,
     external_images: extensions::ExternalImages,
     image_use: ImageUseMap,
@@ -64,18 +66,33 @@ impl fmt::Debug for Surface {
 impl Surface {
     pub fn open(
         gen: ViewProcessGen,
-        mut cfg: HeadlessRequest,
+        cfg: HeadlessRequest,
         window_target: &EventLoopWindowTarget<AppEvent>,
         gl_manager: &mut GlContextManager,
+        mut window_exts: Vec<(ApiExtensionId, Box<dyn WindowExtension>)>,
         mut renderer_exts: Vec<(ApiExtensionId, Box<dyn RendererExtension>)>,
         event_sender: AppEventSender,
     ) -> Self {
         let id = cfg.id;
 
+        for (id, ext) in &mut window_exts {
+            ext.configure(&mut WindowConfigArgs {
+                config: cfg.extensions.iter().find(|(k, _)| k == id).map(|(_, v)| v),
+                window: None,
+            });
+        }
+
         let mut context = gl_manager.create_headless(id, window_target, cfg.render_mode, &event_sender);
         let size = cfg.size.to_px(cfg.scale_factor);
         context.resize(size.to_winit());
-        let context = context;
+
+        window_exts.retain_mut(|(_, ext)| {
+            ext.window_inited(&mut WindowInitedArgs {
+                window: None,
+                context: &mut context,
+            });
+            !ext.is_init_only()
+        });
 
         let mut opts = webrender::WebRenderOptions {
             // text-aa config from Firefox.
@@ -99,13 +116,8 @@ impl Surface {
         };
         let mut blobs = BlobExtensionsImgHandler(vec![]);
         for (id, ext) in &mut renderer_exts {
-            let cfg = cfg
-                .extensions
-                .iter()
-                .position(|(k, _)| k == id)
-                .map(|i| cfg.extensions.swap_remove(i).1);
             ext.configure(&mut RendererConfigArgs {
-                config: cfg,
+                config: cfg.extensions.iter().find(|(k, _)| k == id).map(|(_, v)| v),
                 options: &mut opts,
                 blobs: &mut blobs.0,
             });
@@ -137,9 +149,9 @@ impl Surface {
                 api: &mut api,
                 document_id,
                 pipeline_id,
-                gl: &**context.gl(),
+                context: &mut context,
             });
-            !ext.is_config_only()
+            !ext.is_init_only()
         });
 
         Self {
@@ -153,6 +165,7 @@ impl Surface {
 
             context,
             renderer: Some(renderer),
+            window_exts,
             renderer_exts,
             external_images,
             image_use: ImageUseMap::default(),
@@ -394,7 +407,7 @@ impl Surface {
                 ext.redraw(&mut RedrawArgs {
                     scale_factor: self.scale_factor,
                     size,
-                    gl: &**self.context.gl(),
+                    context: &mut self.context,
                 });
             }
 
@@ -431,6 +444,20 @@ impl Surface {
         images.frame_image(&**self.context.gl(), rect, self.id, self.rendered_frame_id, self.scale_factor, mask)
     }
 
+    /// Calls the window extension command.
+    pub fn window_extension(&mut self, extension_id: ApiExtensionId, request: ApiExtensionPayload) -> ApiExtensionPayload {
+        for (key, ext) in &mut self.window_exts {
+            if *key == extension_id {
+                return ext.command(&mut WindowCommandArgs {
+                    window: None,
+                    context: &mut self.context,
+                    request,
+                });
+            }
+        }
+        ApiExtensionPayload::unknown_extension(extension_id)
+    }
+
     /// Calls the render extension command.
     pub fn render_extension(&mut self, extension_id: ApiExtensionId, request: ApiExtensionPayload) -> ApiExtensionPayload {
         let mut redraw = false;
@@ -455,7 +482,7 @@ impl Surface {
                 ext.redraw(&mut RedrawArgs {
                     scale_factor: self.scale_factor,
                     size,
-                    gl: &**self.context.gl(),
+                    context: &mut self.context,
                 });
             }
         }
@@ -479,8 +506,14 @@ impl Drop for Surface {
             ext.renderer_deinited(&mut RendererDeinitedArgs {
                 document_id: self.document_id,
                 pipeline_id: self.pipeline_id,
-                gl: &**self.context.gl(),
+                context: &mut self.context,
             })
+        }
+        for (_, ext) in &mut self.window_exts {
+            ext.window_deinited(&mut WindowDeinitedArgs {
+                window: None,
+                context: &mut self.context,
+            });
         }
     }
 }